Important Visual Changes!
Added a white transparent DVD logo so the background can be black. (good for OLED screens!)

Also added a walls & corner counter.

The only problem is that the logo has to hit both walls on the EXACT SAME FRAME. Otherwise it's not technically a corner hit. Which might not be satisfying for some users. But technically speaking that's a true corner hit.

Feel free to improve that to add more leniency.

Also one other thing I want to add at some point is the ability to mute the beeping sound through a key press and maybe pause the dvd logo at a specific key press too.

Might also be nice if you could increase the speed of the dvd logo bouncing on a scale of 1 to 10 or something.

Those should be easy to add in but I've not got time right now.